% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_search.R
\name{random_search}
\alias{random_search}
\title{Random Hyperparameter Search}
\usage{
random_search(FUN, bounds, niter = 100, verbose = TRUE)
}
\arguments{
\item{FUN}{Function to be optimised}

\item{bounds}{list object specifying names and ranges of hyperparameters}

\item{niter}{number of random interations}

\item{verbose}{boolean value that controls verbosity}
}
\description{
Random Hyperparameter Search
}
